Speed up sync_fofb.py with wait_for_connection.
This is following Ana's (@anacso17) recommendations for using
wait_for_connection() for all PVs (though we skip the single EVG one
that's the last step of the script).

The consume() function came from [1] as a clean way to loop through all
members of the function with a list comprehension and without wasting
memory.

Since we are here, also make the time_frame_len value a variable that's
set at the top of the script.

Results (reproducible) for runtime (with the sleep(.3) block removed)
were:

Before this commit:

    real    0m55.601s
    user    0m7.838s
    sys     0m0.810s

With this commit:

    real    0m27.843s
    user    0m4.902s
    sys     0m0.676s

[1] https://stackoverflow.com/questions/50937966/fastest-most-pythonic-way-to-consume-an-iterator