Speed up sync_fofb.py even more.
Instead of writing each PV for its crate, we write them as whole lists
where possible. This makes it simpler to have arrays of PVs we can use
in wait_pv() after grouping many writes. CCEnable, specifically, we want
to have always checked and waited for sequentially; the problem that was
happening with it not updating fast enough, however, seems to have been
solved.

The loop in wait_pv() was inspired by [1].

A future direction to go in might be to build a list of PVs and values
as put_pv is called and call a function to consume that list, instead of
having to keep track of what's being waited for and not manually.

This commit brings execution time to under 3s.

[1] https://pyepics.github.io/pyepics/pv.html#pv-connection-callbacks-label