Only enable CC for M1 and M2 BPMs in sync_fofb.py.
It was observed via testing that the max Time of Arrival for packages in
the DCC network was partly determined by the amount of packages going
through the network. Therefore, since the control loop only uses BPMs M1
and M2, we can simply transmit only their packages. This means a
reduction in necessary Time Frame Length for all packages to reach their
recipients from around 3200 to 2000.

Crate 5, unfortunately, has issues in its backplane, so it's necessary
to enable its slots 8 and 10 as well (8 is the only one that's talking
to the FOFB board, 10 is the one that's talking to slot 7, which is what
we are interested in).