sync_fofb: use readback PVs to check values.
Switch from using PV objects and instead use PVPair objects with
setpoint and readback members. The global list still uses individual
PVs, for ease of looping through them.

This required adding a print_pv() function to be used for diagnostics.
For now, we are using the 'sp' field in PVPair for '-Mon' PVs, even if
it isn't exactly coherent; it's working as the "main" field as well.