equalize_bpm_antennas: add script.
This script is used to compute the gains for each (antenna, switching
semicycle) combination and output a list of PVs and the values they
should be set to. It receives acquisition data from a pickle file, and
uses that to determine what it's going to compute; so be sure to only
feed it what you actually want!