sync_fofb: rename to python directory.
The scripts we want to keep here aren't only about syncing the FOFB
network, so they shouldn't be kept in a file hierarchy that makes it
seem so. The helper files are moved to a bpm_apps subdirectory, in order
to keep them separate from the actual executables. Executables won't
need a .py extension, so remove it from sync_fofb.py.