sync_fofb: set up FOFB board trigger lines.
When these triggers are left in their default state (i.e., after an FPGA
reset), they take up the trigger bus and stop the FOFB triggers from
reaching the BPM boards, which is necessary for the synchronization
sequence.

Since we were setting up these boards with a new IOC, it was necessary
to set these values correctly, so, in the future, they would be
retrieved from autosave.

Strictly speaking, it would be more consistent to also configure these
trigger lines for our BPM boards, but we chose not to do that for now.