decode_dcc: add script.
This script is based on [1], with some changes to remove unused modes
and to get its data from an EPICS PV instead of HALCS's acq utility's
output.

The acquisitions must be driven externally, this script is only
responsible for reading the resulting data and interpreting it. It can't
be used for streaming data from the PV either.

[1] https://github.com/lnls-dig/fofb-epics-ioc/blob/master/scripts/python/reformat_fofb_packet_16b.py