update_erics: add script.
This script updates (and installs!) the μHAL based FOFB IOC (and will do
so for future IOCs as well).

It was necessary to mask the halcs-be@3 service on the crates so it
didn't conflict with the FOFB IOC, because we were unable to find what
was launching that halcs-be instance.