acquire_bpm_antennas_for_eq: add script.
Orchestrates the data acquisition for equalizing the antennas. We use an
external trigger and the Monit clock as source so the acquisitions all
start in the same switching semi-cycle.

This script does not restore the trigger mux configuration (maybe it
should?), so it should be re-configured afterwards.

equalize_bpm_antennas has not been updated to use the data from this
acquisition yet.