save_lamp: rename to acquire_lamp.
It now has a few more options for driving the acquisition, and can also
control the power supplies in triggered mode in order to obtain data
from voltage or current steps.

It should be changed to use a named tuple for the current and voltage
fields - as well as possibly spelling them out explicitly - instead of a
dictionary with the 'i' and 'v' keys.