acquire_bpm_antennas_for_eq: change sample amount.
When using any sample amount above 2046, we observed an issue where the
acquisitions for different BPMs (sometimes even on the same board) would
happen in different switching semi-cycles. This was observed by setting
the Dir gain for one of the antennas to 0 - and leaving the Inv gain in
its default value - and plotting that antenna for all BPMs.

While it is unclear what the root of the issue is, we can calibrate the
BPM antenna gains with fewer samples, and that's what we decided to do
here.