update_erics: change runtime user.
Use "iocs", since it's more generic. Also chown recursively so the
autosave files can be overwritten. We also create the user on the
machine, as needed.

Since we are here, also add '--exclude .git' to our rsync command. This
is helpful in avoiding issues when copying from main git repositories
and from worktrees, and eliminates unnecessary network traffic.