update_erics: rename to update_afc_ioc.
This commit also adapts the script to support the changes made in
afc-epics-ioc re. using a single service name for the IOCs.

To avoid impacting operation heavily, we are still using the old
autosave directory. Furthermore, in order to perform this update
properly if erics is still running, the following commands are
necessary:

    rm -f /etc/udev/rules.d/95-erics.rule
    systemctl stop erics-tim@1 erics-fofb@2-1