update_afc_ioc: run for each crate in parallel.
Update operations may take a while to finish, making the overall time
significant when serializing execution. Run each of them in parallel
instead, since they should not influence in one another.

If this is run through a bridge host, it is important to ensure that the
number of pending ssh connection is high enough for all crates.
Otherwise, some of them might be disconnected by the bridge.

Co-authored-by: Érico Nogueira Rolim <erico.rolim@lnls.br>