Add utilitary to disable JTAG access via USB in ftdi device.
This utilitary doesn't work as expected, at this moment, due to a
hardware bug on AFCv4.0 design, but can work in the future when the bugs
are fixed.

The FT4232H (IC21) port A is used as a JTAG interface (pins ADBUS0,
ADBUS1, ADBUS2 and ADBUS3) and pin ADBUS7 is used as a generic GPIO to
enable (logic level high) or disable (logic level low) the level
shifters that connect the FTDI chip to the AMC JTAG bus.

Unfortunately, the transistor logic is wrong and the USB-JTAG level
shifters are always enabled no matter what logic level is set to ADBUS7.